some analyzer error reporting fixes

+++ offset.h
@@ -4,8 +4,8 @@
 
 typedef struct {
   const char *line;
-  int line_length;
-  int column_number;
+  size_t line_length;
+  size_t column_number;
   size_t line_number;
 } OffsetInfo;
